Adding Initial Files to ENSC 57
These files contain everything needed for analyzing our lighting solution. The function "preprocessENSC57" converts video from the cameras to 1-Channel image arrays which is our first step for processing the videos. The other script file is used to compare the amount of light in each image for us to make a decision on which lights to used. In this case we went with the CMVision lights.